+++ source/vue/VueCRS.cpp
@@ -112,7 +112,7 @@ void VueCRS::afficheScreen1(void) {
 	switch (m_crs_screen_mode) {
 	case eVueCRSScreenDataFull:
 	{
-		this->cadran(1, VUE_CRS_NB_LINES, 1, "Dist", _fmkstr(att.dist / 1000., 1U), "km");
+		this->cadran(1, VUE_CRS_NB_LINES, 1, "Dist", _fmkstr(att.dist / 1000.f, 1U), "km");
 		this->cadran(1, VUE_CRS_NB_LINES, 2, "Pwr", _imkstr(att.pwr), "W");
 
 		this->cadran(2, VUE_CRS_NB_LINES, 1, "Speed", _fmkstr(att.loc.speed, 1U), "km/h");
@@ -127,12 +127,12 @@ void VueCRS::afficheScreen1(void) {
 		this->cadranH(5, VUE_CRS_NB_LINES, "Next", _imkstr(att.next), "m");
 
 		float avg_speed = 0.;
-		if (att.nbsec_act) avg_speed = att.dist * 3.6 / att.nbsec_act;
+		if (att.nbsec_act) avg_speed = att.dist * 3.6f / att.nbsec_act;
 		this->cadran(6, VUE_CRS_NB_LINES, 1, "Avg", _fmkstr(avg_speed, 2U), "km/h");
 		this->cadran(6, VUE_CRS_NB_LINES, 2, "Score", _fmkstr(suffer_score.getScore(), 1U), NULL);
 
 		this->cadran(7, VUE_CRS_NB_LINES, 1, "STC", _imkstr((int)stc.getCurrent()), "mA");
-		this->cadran(7, VUE_CRS_NB_LINES, 2, "SOC", _imkstr(percentageBatt(stc.getVoltage(), stc.getCurrent())), "%");
+		this->cadran(7, VUE_CRS_NB_LINES, 2, "SOC", _imkstr((int)percentageBatt(stc.getVoltage(), stc.getCurrent())), "%");
 
 	}  break;
 
@@ -182,7 +182,7 @@ void VueCRS::afficheScreen1(void) {
 			this->cadran(3, VUE_CRS_NB_LINES, 1, "CAD", _imkstr(bsc_info.cadence), "rpm");
 			this->cadran(3, VUE_CRS_NB_LINES, 2, "Climb", _fmkstr(att.climb, 1U), "m");
 
-			this->cadran(4, VUE_CRS_NB_LINES, 1, "Dist", _fmkstr(att.dist / 1000., 1U), "km");
+			this->cadran(4, VUE_CRS_NB_LINES, 1, "Dist", _fmkstr(att.dist / 1000.f, 1U), "km");
 
 			this->cadran(4, VUE_CRS_NB_LINES, 2, "STC", _imkstr((int)stc.getCurrent()), "mA");
 
@@ -233,7 +233,7 @@ void VueCRS::afficheScreen1(void) {
 
 void VueCRS::afficheScreen2(void) {
 
-	this->cadran(1, VUE_CRS_NB_LINES, 1, "Dist", _fmkstr(att.dist / 1000., 1U), "km");
+	this->cadran(1, VUE_CRS_NB_LINES, 1, "Dist", _fmkstr(att.dist / 1000.f, 1U), "km");
 	this->cadran(1, VUE_CRS_NB_LINES, 2, "Pwr", _imkstr(att.pwr), "W");
 
 	this->cadran(2, VUE_CRS_NB_LINES, 1, "Speed", _fmkstr(att.loc.speed, 1U), "km/h");
@@ -298,8 +298,8 @@ void VueCRS::afficheSegment(uint8_t ligne, Segment *p_seg) {
 	ASSERT(liste);
 
 	// compute convertion between deg and meters
-	float deglon_to_m = 1000. * distance_between(att.loc.lat, att.loc.lon, att.loc.lat, att.loc.lon + 0.001);
-	float deglat_to_m = 1000. * distance_between(att.loc.lat, att.loc.lon, att.loc.lat + 0.001, att.loc.lon);
+	float deglon_to_m = 1000.f * distance_between(att.loc.lat, att.loc.lon, att.loc.lat, att.loc.lon + 0.001f);
+	float deglat_to_m = 1000.f * distance_between(att.loc.lat, att.loc.lon, att.loc.lat + 0.001f, att.loc.lon);
 
 	if (p_seg->getStatus() == SEG_OFF) {
 
@@ -311,12 +311,12 @@ void VueCRS::afficheSegment(uint8_t ligne, Segment *p_seg) {
 		LOG_DEBUG("Printing SEG OFF\r\n");
 
 		// zoom level
-		minLat -= 300 / deglat_to_m;
-		minLon -= 300 / deglon_to_m;
-		minAlt -= 15.;
-		maxLat += 300 / deglat_to_m;
-		maxLon += 300 / deglon_to_m;
-		maxAlt += 15.;
+		minLat -= 300.f / deglat_to_m;
+		minLon -= 300.f / deglon_to_m;
+		minAlt -= 15.f;
+		maxLat += 300.f / deglat_to_m;
+		maxLon += 300.f / deglon_to_m;
+		maxAlt += 15.f;
 
 	} else {
 
@@ -349,10 +349,10 @@ void VueCRS::afficheSegment(uint8_t ligne, Segment *p_seg) {
 		// marge
 		minLat -= 100 / deglat_to_m;
 		minLon -= 100 / deglon_to_m;
-		minAlt -= 15.;
+		minAlt -= 15.f;
 		maxLat += 100 / deglat_to_m;
 		maxLon += 100 / deglon_to_m;
-		maxAlt += 15.;
+		maxAlt += 15.f;
 
 		LOG_DEBUG("Printing SEG status=%d\r\n", p_seg->getStatus());
 
@@ -387,9 +387,9 @@ void VueCRS::afficheSegment(uint8_t ligne, Segment *p_seg) {
 
 			if (!pSuivant.isValid() || !pCourant.isValid()) break;
 
-			drawLine(regFenLim(pCourant._lon, minLon, maxLon, 0, _width),
+			drawLine(regFenLim(pCourant._lon, minLon, maxLon, 0.f, _width),
 					regFenLim(pCourant._lat, minLat, maxLat, fin_cadran, debut_cadran),
-					regFenLim(pSuivant._lon, minLon, maxLon, 0, _width),
+					regFenLim(pSuivant._lon, minLon, maxLon, 0.f, _width),
 					regFenLim(pSuivant._lat, minLat, maxLat, fin_cadran, debut_cadran), LS027_PIXEL_BLACK);
 		}
 
@@ -400,12 +400,12 @@ void VueCRS::afficheSegment(uint8_t ligne, Segment *p_seg) {
 
 	// draw a circle at the end of the segment
 	if (p_seg->getStatus() < SEG_OFF) {
-		drawCircle(regFenLim(pSuivant._lon, minLon, maxLon, 0, _width),
+		drawCircle(regFenLim(pSuivant._lon, minLon, maxLon, 0.f, _width),
 				regFenLim(pSuivant._lat, minLat, maxLat, fin_cadran, debut_cadran), 5, LS027_PIXEL_BLACK);
 	} else if (p_seg->getStatus() > SEG_OFF) {
 		// draw a circle at the start of the segment
 		maPos = liste->getFirstPoint();
-		drawCircle(regFenLim(maPos->_lon, minLon, maxLon, 0, _width),
+		drawCircle(regFenLim(maPos->_lon, minLon, maxLon, 0.f, _width),
 				regFenLim(maPos->_lat, minLat, maxLat, fin_cadran, debut_cadran), 5, LS027_PIXEL_BLACK);
 	}
 
@@ -426,22 +426,22 @@ void VueCRS::afficheSegment(uint8_t ligne, Segment *p_seg) {
 	maDpex = regFenLim(_lon, minLon, maxLon, 0, _width);
 	maDpey = regFenLim(_lat, minLat, maxLat, fin_cadran, debut_cadran);
 	if (att.loc.course > 0) {
-		int16_t x0f, x0 = maDpex;
-		int16_t y0f, y0 = maDpey - 15;
+		int16_t x0f, x0 = (int16_t)maDpex;
+		int16_t y0f, y0 = (int16_t)maDpey - 15;
 
-		int16_t x1f, x1 = maDpex + 5;
-		int16_t y1f, y1 = maDpey + 5;
+		int16_t x1f, x1 = (int16_t)maDpex + 5;
+		int16_t y1f, y1 = (int16_t)maDpey + 5;
 
-		int16_t x2f, x2 = maDpex - 5;
-		int16_t y2f, y2 = maDpey + 5;
+		int16_t x2f, x2 = (int16_t)maDpex - 5;
+		int16_t y2f, y2 = (int16_t)maDpey + 5;
 
-		rotate_point(att.loc.course, maDpex, maDpey, x0, y0, x0f, y0f);
-		rotate_point(att.loc.course, maDpex, maDpey, x1, y1, x1f, y1f);
-		rotate_point(att.loc.course, maDpex, maDpey, x2, y2, x2f, y2f);
+		rotate_point(att.loc.course, (int16_t)maDpex, (int16_t)maDpey, x0, y0, x0f, y0f);
+		rotate_point(att.loc.course, (int16_t)maDpex, (int16_t)maDpey, x1, y1, x1f, y1f);
+		rotate_point(att.loc.course, (int16_t)maDpex, (int16_t)maDpey, x2, y2, x2f, y2f);
 
 		drawTriangle(x0f, y0f, x1f, y1f, x2f, y2f, LS027_PIXEL_BLACK);
 	} else {
-		fillCircle(maDpex, maDpey, 4, LS027_PIXEL_BLACK);
+		fillCircle((int16_t)maDpex, (int16_t)maDpey, 4, LS027_PIXEL_BLACK);
 	}
 
 	// return before printing text
@@ -450,9 +450,9 @@ void VueCRS::afficheSegment(uint8_t ligne, Segment *p_seg) {
 	}
 
 	if (maDpey > fin_cadran - 30) {
-		setCursor(maDpex > _width - 70 ? _width - 70 : maDpex, maDpey - 20);
+		setCursor(maDpex > _width - 70 ? _width - 70 : (int16_t)maDpex, (int16_t)maDpey - 20);
 	} else {
-		setCursor(maDpex > _width - 70 ? _width - 70 : maDpex, maDpey + 15);
+		setCursor(maDpex > _width - 70 ? _width - 70 : (int16_t)maDpex, (int16_t)maDpey + 15);
 	}
 
 	setTextSize(2);
@@ -490,7 +490,7 @@ void VueCRS::partner(uint8_t ligne, Segment *p_seg) {
 	curtime = p_seg->getCur();
 
 	if (curtime < 5.) {
-		indice = rtime / 5.;
+		indice = rtime / 5.f;
 	} else {
 		indice = rtime / curtime;
 	}
@@ -504,21 +504,21 @@ void VueCRS::partner(uint8_t ligne, Segment *p_seg) {
 	// -> si -0.25 centre en ol
 	// -> si  0.25 centre en 240-ol
 	//centre = ol + (LCDWIDTH - ol - ol) * (0.25 + indice) / (0.5);
-	centre = regFenLim(indice, -0.25, 0.25, ol, _width - ol);
+	centre = regFenLim(indice, -0.25f, 0.25f, ol, _width - ol);
 	if (centre < ol) {
 		centre = ol;
 	} else if (centre > (int)_width - ol) {
 		centre = _width - ol;
 	}
 
-	dixP = (_width - 2.*ol) * 10. / 50.;
+	dixP = (_width - 2 * ol) * 10 / 50;
 	drawFastVLine(_width / 2 - dixP, hl + 2, 7, LS027_PIXEL_BLACK);
 	drawFastVLine(_width / 2 + dixP, hl + 2, 7, LS027_PIXEL_BLACK);
 
 	fillTriangle(centre - 7, hl + 7, centre, hl - 7, centre + 7, hl + 7, LS027_PIXEL_BLACK);
 	setCursor(centre - 15, hl + 12);
 	setTextSize(VUE_CRS_NB_LINES > 7 ? 1 : 2);
-	print(_imkstr((int)(indice * 100.)));
+	print(_imkstr((int)(indice * 100.f)));
 
 	// marques
 	drawFastVLine(_width / 2, hl - 12, 12, LS027_PIXEL_BLACK);
